Post-move cleanup
