+++ pom.xml
@@ -38,6 +38,8 @@
     <revision>1.3.99999-SNAPSHOT</revision>
     <junit.jupiter.version>5.7.0-M1</junit.jupiter.version>
     <testcontainers.version>1.14.2</testcontainers.version>
+    <hamcrest.version>2.2</hamcrest.version>
+    <mockito.version>1.10.19</mockito.version>
   </properties>
 
   <modules>
@@ -129,13 +131,13 @@
       <dependency>
         <groupId>org.mockito</groupId>
         <artifactId>mockito-all</artifactId>
-        <version>1.10.19</version>
+        <version>${mockito.version}</version>
         <scope>test</scope>
       </dependency>
       <dependency>
         <groupId>org.hamcrest</groupId>
         <artifactId>hamcrest-core</artifactId>
-        <version>2.2</version>
+        <version>${hamcrest.version}</version>
         <scope>test</scope>
       </dependency>
       <dependency>
