+++ transactionoutbox-core/pom.xml
@@ -47,15 +47,14 @@
       <artifactId>javax.el</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>com.ea.async</groupId>
-      <artifactId>ea-async</artifactId>
-    </dependency>
-
     <!-- Compile time -->
     <dependency>
       <groupId>org.projectlombok</groupId>
       <artifactId>lombok</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.ea.async</groupId>
+      <artifactId>ea-async</artifactId>
     </dependency>
 
     <!-- Test dependencies -->
