+++ transactionoutbox-core/src/test/java/com/gruelbox/transactionoutbox/acceptance/TestComplexConfigurationExample.java
