+++ transactionoutbox-testing/pom.xml
@@ -38,13 +38,13 @@
       <dependency>
         <groupId>org.mockito</groupId>
         <artifactId>mockito-all</artifactId>
-        <version>1.10.19</version>
+        <version>${mockito.version}</version>
         <scope>compile</scope>
       </dependency>
       <dependency>
         <groupId>org.hamcrest</groupId>
         <artifactId>hamcrest-core</artifactId>
-        <version>2.2</version>
+        <version>${hamcrest.version}</version>
         <scope>compile</scope>
       </dependency>
     </dependencies>
